pulseplane_backend/requirements.txt
Django
django-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
dj-database-url
whitenoise
psycopg2-binary
gunicorn
